Edit File OK

+++ src/app/api/document/other/[id]/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from "next/server";
 import { prisma } from "@/libs/prisma";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
+import path from 'path';
+import { promises as fs } from 'fs';
 
 interface ParamsProps {
   params: {
@@ -8,50 +10,38 @@ interface ParamsProps {
   };
 }
 
-interface BodyProps {
-    name?:string,
-    depar?:string,
-    assignedToId?: string;
-}
 
 export async function DELETE(req: Request, { params }: ParamsProps) {
-    const { id } = params;
-    try {
-      const file = await prisma.file.delete({ where: { id } });
-      return NextResponse.json(file);
-    } catch (error) {
-      if (error instanceof PrismaClientKnownRequestError && error.code === 'P2025') {
-        return NextResponse.json({ message: "Archivo no encontrado" }, { status: 400 });
-      } else {
-        console.error('Error eliminando archivo:', error);
-        return NextResponse.json({ message: "Error Interno" }, { status: 500 });
-      }
-    }
-  }
+  const { id } = params;
 
-  export async function PUT(req: Request, { params }: ParamsProps) {
-    const { id } = params;
-    const body: BodyProps = await req.json();
-  
-    try {
-      // Actualizar el archivo con los nuevos datos
-      const updatedFile = await prisma.file.update({
-        where: { id },
-        data: {
-          name: body.name,
-          depar: body.depar,
-          assignedToId: body.assignedToId,
-        },
-      });
-  
-      // Devolver la respuesta con el archivo actualizado
-      return NextResponse.json(updatedFile);
-    } catch (error) {
-      if (error instanceof PrismaClientKnownRequestError && error.code === 'P2025') {
-        return NextResponse.json({ message: "Archivo no encontrado" }, { status: 404 });
-      } else {
-        console.error('Error actualizando archivo:', error);
-        return NextResponse.json({ message: "Error Interno" }, { status: 500 });
-      }
+  try {
+    // Primero, obtén la información del archivo a eliminar
+    const file = await prisma.file.findUnique({ where: { id } });
+
+    if (!file) {
+      return NextResponse.json({ message: "Archivo no encontrado" }, { status: 400 });
     }
+
+    // Elimina la entrada del archivo en la base de datos
+    await prisma.file.delete({ where: { id } });
+
+    // Construye la ruta completa del archivo en el sistema de archivos
+    const filePath = path.join(process.cwd(), 'public', file.path);
+
+    // Intenta eliminar el archivo del sistema de archivos
+    await fs.unlink(filePath);
+
+    return NextResponse.json(file);
+  } catch (error) {
+    if (error instanceof PrismaClientKnownRequestError && error.code === 'P2025') {
+      return NextResponse.json({ message: "Archivo no encontrado" }, { status: 400 });
+    } else {
+      console.error('Error eliminando archivo:', error);
+      return NextResponse.json({ message: "Error Interno" }, { status: 500 });
+    }
+  } finally {
+    await prisma.$disconnect();
   }
+}
+
+
